Fixed Dropbox authentication
Dropbox wouldn't properly authenticate when using the redirection method. In
order to get it to behave properly, we had to completely change the authentication driver
and instead use the popup method which opens a separate window. This method works properly on production setups and doesn't conflict with Angular's usage of
routes.