ItemMirror Angular Integration
Deterimined the problem with Angular not propogating any changes to
the associations (e.g. a deleted association doesn't disappear). The core
of the problem lies with how Angular handles scope and updates. Because
we're primarly using a service to pass data around, we also need to make
sure that our scope is updated, otherwise Angular has no way to keep track
of the changes that we make.

The simple solution is to wrap all promises with a scope update function,
which modifies the associations so that they match that of the service.

Currently, this application is capable of displaying any data that an association has
as well as deleting associations and entering folders. With the scope and display isses solved, it's now simple to start building a complex CRUD application.