Changed default host settings; removed needless file
Modified the host settings under the Gruntfile so that localhost is now the default, rather than my personal domain. The only change that needs to be made is in dropbox-auth.js, since this is the dropbox redirect.

Also updated the welcome page to inform people that popups must be enabled for the app to function.