Added a namespace attribute represented as JSON
Added a 'tags', attribute, which allows for simple addition and subtraction of tags associated to assiociations Internally this is represented as JSON, and it has a set of public methods to modify this object.

This is a huge step, because it allows for storage of much more complex data in a single attribute, and easy manipulation of that data through a self imposed set of methods. For applications which move beyond the need of simple read write bvalues, this will come in handy.